cahtbot/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sentence-transformers==2.3.1
httpx==0.26.0
slowapi==0.1.9
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.8
